refactor: capitalize component directory for naming convention
